Alternative memory backend map based on std HashMap
dashmap documentation states it can deadlock in concurrent scenarios, for instance:
https://docs.rs/dashmap/5.3.4/dashmap/struct.DashMap.html#method.entry

Recent versions seem to have reverted to hand-written spinlocks, which resolve the
deadlock situation but might result in poor performance and elevated CPU load:
https://github.com/xacrimon/dashmap/issues/206

Given the lack of support for entry APIs on the majority of the concurrent HashMap
implementations and lingering possibility of deadlocks on some of them depending on
how the map is accessed, this commit introduces a more conservative approach that
might be a better fit for general scenarios: a simple HashMap behind a Mutex.

Performance-wise, it should operate well on low to moderate load and will not
deadlock regardless of access order since there is a single lock controlling
the access of the whole map.